continue

+++ Interfaces/IMWPlayerInfo.cs
@@ -9,7 +9,7 @@ namespace MultiWorldLib.Interfaces
 {
     public interface IMWPlayerInfo
     {
-        public int Index { get; set; }
+        public ushort Index { get; set; }
         public int SpawnX { get; set; }
         public int SpawnY { get; set; }
     }
